Unset/set proxy when activating a profile.
Long-lived flows are not affected if we change a profile during their lifetime.
Their cookie has already been set on ATCD, and a new cookie won't be detected.
By resetting the proxy we force them to go first through their destination and
then through the proxy, so we can attach the cookie on the newly created flow
and then detect it on the network.