Unproxy HTTPS for online version.
HTTPS comes with half-baked support because it requires a custom Chromium to
attach cookies to the TLS handshake.
For standard Chrome, we could proxy https anyways, and treat them as best-effort traffic
(since cookies are not visible).
Instead, we opt to send only cookie-enabled traffic through the proxy.
The main reason is to use the cookie as an authenticaion primitive, besides picking a QoS profile.
Anytime I setup a proxy, a bot detects it within a few hours, and traffic spikes. To avoid this,
only cookie-enabled traffic can make it through the proxy.